pyproject.toml
[project]
name = "llm-service"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "cerebras-cloud-sdk>=1.0",
    "pydantic-settings>=2.0",
]
